Add "Power" tag to more cards.
I'm assuming that this tag is meant for cards that have an icon cost to use its power. In that case the following cards were missing that tag:
Convoys
Fae Merchant
War Tortoise

The name of this tag is a bit confusing since power is a much broader term in Oath. Maybe something like "Cost" would be more fitting?